Add files via upload
This Data Analysis project analyses data relating to customers application for bank loans. 

The data was initially uploaded and cleaned in Excel and transferred to SQL as a CSV file. I explored the data further in SQL to analyse and highlight key KPIs. I used functions such as JOINs and GROUP BY to undertake analysis. 

I used Tableau and created a interactive dashboard to visualise the data clearly, highlighting key KPIs to clearly communicate insights for stakeholders.